services/template-service/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=services/template-service/requirements.txt services/template-service/requirements.in
#
aenum==3.1.16
    # via gremlinpython
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.12.14
    # via gremlinpython
aiosignal==1.4.0
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==3.7.1
    # via
    #   fastapi
    #   starlette
    #   watchfiles
asgiref==3.9.1
    # via opentelemetry-instrumentation-asgi
attrs==25.3.0
    # via aiohttp
authlib==1.2.1
    # via -r /Users/m/Downloads/platformQ/requirements/base.in
avro==1.11.3
    # via -r /Users/m/Downloads/platformQ/requirements/base.in
backoff==2.2.1
    # via
    #   opentelemetry-exporter-otlp-proto-common
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
bcrypt==4.3.0
    # via passlib
black==25.1.0
    # via -r services/template-service/requirements.in
cachetools==5.5.2
    # via google-auth
cassandra-driver==3.28.0
    # via -r /Users/m/Downloads/platformQ/requirements/base.in
certifi==2025.7.14
    # via
    #   kubernetes
    #   pulsar-client
    #   requests
cffi==1.17.1
    # via cryptography
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via
    #   black
    #   geomet
    #   uvicorn
cryptography==45.0.5
    # via
    #   authlib
    #   python-jose
deprecated==1.2.18
    # via
    #   opentelemetry-api
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
ecdsa==0.19.1
    # via python-jose
fastapi==0.104.1
    # via -r /Users/m/Downloads/platformQ/requirements/base.in
frozenlist==1.7.0
    # via
    #   aiohttp
    #   aiosignal
geomet==0.2.1.post1
    # via cassandra-driver
gitdb==4.0.12
    # via gitpython
gitpython==3.1.44
    # via -r services/template-service/requirements.in
google-auth==2.40.3
    # via kubernetes
googleapis-common-protos==1.70.0
    # via
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
gremlinpython==3.7.0
    # via -r /Users/m/Downloads/platformQ/requirements/base.in
grpcio==1.59.2
    # via
    #   -r /Users/m/Downloads/platformQ/requirements/base.in
    #   grpcio-tools
    #   opentelemetry-exporter-otlp-proto-grpc
grpcio-tools==1.59.2
    # via -r /Users/m/Downloads/platformQ/requirements/base.in
h11==0.16.0
    # via uvicorn
httptools==0.6.4
    # via uvicorn
hvac==1.2.1
    # via -r /Users/m/Downloads/platformQ/requirements/base.in
idna==3.10
    # via
    #   anyio
    #   requests
    #   yarl
importlib-metadata==6.11.0
    # via opentelemetry-api
iniconfig==2.1.0
    # via pytest
isodate==0.7.2
    # via gremlinpython
kubernetes==28.1.0
    # via -r /Users/m/Downloads/platformQ/requirements/base.in
multidict==6.6.3
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.1.0
    # via black
nest-asyncio==1.6.0
    # via gremlinpython
oauthlib==3.3.1
    # via
    #   kubernetes
    #   requests-oauthlib
opentelemetry-api==1.21.0
    # via
    #   -r /Users/m/Downloads/platformQ/requirements/base.in
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-asgi
    #   opentelemetry-instrumentation-fastapi
    #   opentelemetry-instrumentation-grpc
    #   opentelemetry-sdk
opentelemetry-exporter-otlp==1.21.0
    # via -r /Users/m/Downloads/platformQ/requirements/base.in
opentelemetry-exporter-otlp-proto-common==1.21.0
    # via
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
opentelemetry-exporter-otlp-proto-grpc==1.21.0
    # via opentelemetry-exporter-otlp
opentelemetry-exporter-otlp-proto-http==1.21.0
    # via opentelemetry-exporter-otlp
opentelemetry-instrumentation==0.42b0
    # via
    #   opentelemetry-instrumentation-asgi
    #   opentelemetry-instrumentation-fastapi
    #   opentelemetry-instrumentation-grpc
opentelemetry-instrumentation-asgi==0.42b0
    # via opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-fastapi==0.42b0
    # via -r /Users/m/Downloads/platformQ/requirements/base.in
opentelemetry-instrumentation-grpc==0.42b0
    # via -r /Users/m/Downloads/platformQ/requirements/base.in
opentelemetry-proto==1.21.0
    # via
    #   opentelemetry-exporter-otlp-proto-common
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
opentelemetry-sdk==1.21.0
    # via
    #   -r /Users/m/Downloads/platformQ/requirements/base.in
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-exporter-otlp-proto-http
    #   opentelemetry-instrumentation-grpc
opentelemetry-semantic-conventions==0.42b0
    # via
    #   opentelemetry-instrumentation-asgi
    #   opentelemetry-instrumentation-fastapi
    #   opentelemetry-instrumentation-grpc
    #   opentelemetry-sdk
opentelemetry-util-http==0.42b0
    # via
    #   opentelemetry-instrumentation-asgi
    #   opentelemetry-instrumentation-fastapi
packaging==25.0
    # via
    #   black
    #   pytest
passlib[bcrypt]==1.7.4
    # via -r /Users/m/Downloads/platformQ/requirements/base.in
pathspec==0.12.1
    # via black
platformdirs==4.3.8
    # via black
pluggy==1.6.0
    # via pytest
propcache==0.3.2
    # via
    #   aiohttp
    #   yarl
protobuf==4.25.1
    # via
    #   -r /Users/m/Downloads/platformQ/requirements/base.in
    #   googleapis-common-protos
    #   grpcio-tools
    #   opentelemetry-proto
pulsar-client==3.3.0
    # via -r /Users/m/Downloads/platformQ/requirements/base.in
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   python-jose
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pycparser==2.22
    # via cffi
pydantic==2.5.2
    # via
    #   -r /Users/m/Downloads/platformQ/requirements/base.in
    #   fastapi
pydantic-core==2.14.5
    # via pydantic
pygments==2.19.2
    # via pytest
pyhcl==0.4.5
    # via hvac
pytest==8.4.1
    # via -r services/template-service/requirements.in
python-consul==1.1.0
    # via -r /Users/m/Downloads/platformQ/requirements/base.in
python-dateutil==2.9.0.post0
    # via kubernetes
python-dotenv==1.0.0
    # via
    #   -r /Users/m/Downloads/platformQ/requirements/base.in
    #   uvicorn
python-jose[cryptography]==3.3.0
    # via -r /Users/m/Downloads/platformQ/requirements/base.in
pyyaml==6.0.2
    # via
    #   kubernetes
    #   uvicorn
requests==2.31.0
    # via
    #   -r /Users/m/Downloads/platformQ/requirements/base.in
    #   hvac
    #   kubernetes
    #   opentelemetry-exporter-otlp-proto-http
    #   python-consul
    #   requests-oauthlib
requests-oauthlib==2.0.0
    # via kubernetes
rsa==4.9.1
    # via
    #   google-auth
    #   python-jose
ruff==0.12.3
    # via -r services/template-service/requirements.in
six==1.17.0
    # via
    #   cassandra-driver
    #   ecdsa
    #   geomet
    #   gremlinpython
    #   kubernetes
    #   python-consul
    #   python-dateutil
smmap==5.0.2
    # via gitdb
sniffio==1.3.1
    # via anyio
starlette==0.27.0
    # via fastapi
typing-extensions==4.14.1
    # via
    #   aiosignal
    #   fastapi
    #   opentelemetry-sdk
    #   pydantic
    #   pydantic-core
urllib3==1.26.20
    # via
    #   kubernetes
    #   requests
uvicorn[standard]==0.24.0.post1
    # via -r /Users/m/Downloads/platformQ/requirements/base.in
uvloop==0.21.0
    # via uvicorn
watchfiles==1.1.0
    # via uvicorn
websocket-client==1.8.0
    # via kubernetes
websockets==15.0.1
    # via uvicorn
wrapt==1.17.2
    # via
    #   deprecated
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-grpc
yarl==1.20.1
    # via aiohttp
zipp==3.23.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
